fix(airflow): update connection IDs in DAGs for crypto pipeline
Replaces the connection IDs in the `crypto_pipeline_monitor_dag.py` and `crypto_pipeline_submit_dag.py` files to use specific connections for the crypto pipeline. The `conn_id` for the SQLCheckOperator is changed to `crypto_pipeline_postgres`, and the `conn_id` for the SparkSubmitOperator is updated to `crypto_pipeline_spark`, ensuring proper integration with the respective services.